[issue 16] Close Issue 16 by upgrading jshint to 1.1.0
This commit upgrades the project to use jshint version 1.1.0. In this
version, they've allowed the use of a space before the global keyword.

I've added a unit test that demonstrated this behavior before the
change to the package.json file. I hope it'll be a useful template for
use in other issue testing.