Update to use escomplex-js 1.0.4 for analysis
The `complexity-report` project is no longer recommended for direct use
and has been refactored into components. This updates Plato to use the
[escomplex-js](https://github.com/philbooth/escomplex) directly.

Please see [this issue](https://github.com/philbooth/complexity-report/issues/32)
for more information about the change.

The `escomplex-js` module does seem to have a different data format, most notably
removing a level of the report and moving `sloc` and such up to the top level.
This required a fair number of changes throughout.