Update to use escomplex-js for analysis
The `complexity-report` project is no longer recommended for direct use
and has been refactored into components. This updates Plato to use the
[escomplex-js](https://github.com/philbooth/escomplex) directly.

Please see [this issue](https://github.com/philbooth/complexity-report/issues/32)
for more information about the change.

This also moves some fixture data out of a string an into a fixture file,
it was easier to make this change while having it to visually inspect.